Added stdout and stderr to check_call
It seems, that with long pandoc-files, something happens with stdout and we get the error `BlockingIOError: [Errno 11] write could not complete without blocking`. To avoid this, we can just ignore stderr and stdout from the mermaid call, as they are not used in any case. After adding this parameters, the error vanishes.